refactor: adding components folder in test
