+++ test/shared/components/atoms/DialogContainer.test.jsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import { render, screen, fireEvent, cleanup } from "@testing-library/react";
 import React from "react";
-import { DialogContainer } from "../../../src/shared/components/atoms/DialogContainer";
+import { DialogContainer } from "../../../../src/shared/components/atoms/DialogContainer";
 
 describe("DialogContainer", () => {
   beforeEach(() => {
